refactor cobol doc generator into new source file
